restframework/requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
algoliasearch-django>=4.0,<5.0
